Move collection type detection to a separate module
