+++ default/python/stub_generator/stub_generator/rtype/_function_mapping.py
@@ -1,5 +1,5 @@
-from stub_generator.stub_generator.rtype._base_rtype_mapping import make_type
 from common.fo_typing import EmpireId, Turn
+from stub_generator.stub_generator.coolection_classes import make_type
 
 _rtypes_map = {
     "currentTurn": Turn.__name__,
