+++ default/python/stub_generator/stub_generator/rtype/_property_mapping.py
@@ -1,4 +1,4 @@
-from stub_generator.stub_generator.rtype._base_rtype_mapping import make_type
+from stub_generator.stub_generator.coolection_classes import make_type
 
 
 _property_map = {
